separate the DicomParsarUtils testing into unit and integration testing
